Move shared socket types to shared package
