+++ client/package.json
@@ -17,6 +17,7 @@
     "react-dom": "^18.2.0",
     "react-onsenui": "^1.13.3",
     "react-redux": "^9.1.0",
+    "socket.io-client": "^4.7.4",
     "zod": "^3.22.4"
   },
   "devDependencies": {
