+++ server/src/index.ts
@@ -2,7 +2,8 @@ import { DefaultEventsMap } from 'socket.io/dist/typed-events';
 import { createServer } from 'http';
 import { Server } from 'socket.io';
 import { generalLogger } from './utils/loggers';
-import { ServerToClientEvents, SocketWithAuth, ClientToServerEvents } from './sockets/socket.types';
+import { SocketWithAuth } from './sockets/types';
+import {ClientToServerEvents, ServerToClientEvents} from "shared";
 import { verifyToken } from './sockets/middlewares/verifyToken';
 import onConnection from './sockets';
 import app from './app';
