+++ server/src/sockets/types.ts
@@ -1,0 +1,5 @@
+import { RejoinPollFields } from "../api/polls/types";
+
+export interface SocketWithAuth extends RejoinPollFields {};
+
+export type NextFunction = (err?: any) => void;
