+++ shared/index.ts
@@ -1,1 +1,2 @@
 export * from './poll.types';
+export * from './socket.types';
